+++ doc/references/release-notes.rst
@@ -23,6 +23,11 @@ Upcoming Release
   The new modularized maps module allows for more flexibility and easier
   maintenance. 
 
+* Added new single node capacity expansion example in style of model.energy.
+  It can be loaded with ``pypsa.examples.model_energy()``.
+
+* Add new example for how to run MGA ('modelling-to-generate-alternatives') optimisation.
+
 `v0.34.0 <https://github.com/PyPSA/PyPSA/releases/tag/v0.34.0>`__ (25th March 2025)
 =======================================================================================